@@ -7,19 +7,15 @@
   "author": "Flavien Rebattet",
   "license": "MIT",
   "dependencies": {
-    "@hookform/resolvers": "^2.8.8",
-    "axios": "^0.21.1",
-    "connected-react-router": "^6.9.2",
-    "history": "4.10.1",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-hook-form": "^7.29.0",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "redux": "4.1.1",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
-    "typescript": "^4.4.4",
+    "@hookform/resolvers": "^2.9.7",
+    "axios": "^0.27.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.34.1",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^6.3.0",
+    "redux": "4.2.0",
+    "redux-thunk": "^2.4.1",
     "yup": "^0.32.11"
   },
   "scripts": {
@@ -39,14 +35,11 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@types/history": "^4.7.9",
-    "@types/jest": "^27.4.1",
-    "@types/node": "16.11.26",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^18.0.0",
-    "@types/react-router": "^5.1.18",
-    "@types/react-router-dom": "^5.3.2",
-    "@types/redux-logger": "^3.0.9",
-    "react-scripts": "^5.0.0"
+    "@types/node": "^16.11.48",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.7.4"
   }
 }
