@@ -13,44 +13,44 @@
     "ci": "yarn install --frozen-lockfile"
   },
   "dependencies": {
-    "@sendgrid/mail": "^7.4.6",
-    "@types/bcryptjs": "^2.4.2",
+    "@sendgrid/mail": "^7.7.0",
     "bcryptjs": "^2.4.3",
-    "body-parser": "^1.19.0",
+    "body-parser": "^1.20.0",
     "compression": "^1.7.4",
-    "connect-mongo": "^4.4.1",
+    "connect-mongo": "^4.6.0",
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
-    "dayjs": "^1.10.7",
-    "express": "^4.17.1",
+    "dayjs": "^1.11.5",
+    "express": "^4.18.1",
     "express-async-errors": "^3.1.1",
-    "express-rate-limit": "^6.3.0",
-    "express-session": "^1.17.2",
-    "helmet": "^4.6.0",
-    "joi": "^17.4.2",
-    "joi-objectid": "^3.0.1",
+    "express-rate-limit": "^6.5.1",
+    "express-session": "^1.17.3",
+    "helmet": "^5.1.1",
+    "joi": "^17.6.0",
+    "joi-objectid": "^4.0.2",
     "moment": "^2.29.1",
     "mongo-sanitize": "^1.1.0",
-    "mongoose": "^6.0.12",
-    "passport": "^0.4.1",
+    "mongoose": "^6.5.2",
+    "passport": "^0.6.0",
     "passport-local": "^1.0.0",
-    "ramda": "^0.27.1",
-    "tsc-alias": "^1.6.6",
-    "winston": "^3.3.3"
+    "ramda": "^0.28.0",
+    "winston": "^3.8.1"
   },
   "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
     "@types/compression": "^1.7.2",
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
-    "@types/express-session": "^1.17.4",
+    "@types/express-session": "^1.17.5",
     "@types/mongo-sanitize": "^1.0.1",
-    "@types/node": "^16.11.7",
-    "@types/passport": "^1.0.7",
+    "@types/node": "16.11.48",
+    "@types/passport": "^1.0.10",
     "@types/passport-local": "^1.0.34",
-    "@types/ramda": "^0.27.61",
+    "@types/ramda": "^0.28.15",
     "@types/winston": "^2.4.4",
-    "ts-node-dev": "^1.1.8",
-    "tsconfig-paths": "^4.0.0",
-    "typescript": "^4.4.4"
+    "ts-node-dev": "^2.0.0",
+    "tsc-alias": "^1.7.0",
+    "tsconfig-paths": "^4.1.0",
+    "typescript": "^4.7.4"
   }
 }
