@@ -1,19 +1,18 @@
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
-import { ConnectedRouter } from "connected-react-router";
-import { store, history } from "./store/store";
+import { store } from "./store/store";
 import App from "./App";
 import registerServiceWorker from "./registerServiceWorker";
+import { BrowserRouter } from "react-router-dom";
 
 import "./index.css";
 
 createRoot(document.getElementById("root")!).render(
-  <Provider store={store}>
-    {/* @ts-ignore outdated lib types*/}
-    <ConnectedRouter history={history}>
+  <BrowserRouter>
+    <Provider store={store}>
       <App />
-    </ConnectedRouter>
-  </Provider>
+    </Provider>
+  </BrowserRouter>
 );
 
 registerServiceWorker();
