@@ -1,4 +1,4 @@
-import { useParams } from "react-router-dom";
+import { Navigate, useNavigate, useParams } from "react-router-dom";
 import { attemptGetConfirmation } from "../store/thunks/auth";
 import { Error } from "../components";
 import { useAppDispatch } from "src/store/hooks";
@@ -6,12 +6,17 @@ import { useServerError } from "src/hooks/useServerError";
 
 export default function RegisterConfirmationPage() {
   const dispatch = useAppDispatch();
+  const navigate = useNavigate();
   const { serverError, handleServerError } = useServerError();
 
   const { token } = useParams<{ token: string }>();
 
+  if (!token) {
+    return <Navigate to='/home' replace />;
+  }
+
   const handleSubmit = () => {
-    dispatch(attemptGetConfirmation(token)).catch(handleServerError);
+    dispatch(attemptGetConfirmation(token, navigate)).catch(handleServerError);
   };
 
   return (
